Move component breakdown to it's own page, and same timescale as overall metrics
