+++ client/beacon.js
@@ -289,6 +289,10 @@ Keen.ready(function(){
 			require('./queries/front-page/ctr').render();
 			break;
 
+		case '/graph/front-page/components':
+			require('./queries/front-page/ctr-breakdown').render();
+			break;
+
 		case '/graph/front-page/visits':
 			require('./queries/front-page/visits').render();
 			break;
