+++ client/queries/front-page/ctr/cohorts.js
@@ -23,7 +23,12 @@ const render = (el, promiseOfData) => {
         .prepare();
 
     promiseOfData
-    .then(([, usersByDay, , , clicksByUserAndDay, ]) => {
+    .then((
+        [ , //users
+        usersByDay,
+        clicksByUserAndDay,
+         , //viewsByDay
+        ]) => {
 
         const cohortResults = clicksByUserAndDay.map((result, day) => {
             const newValue = new Array(0, 0, 0, 0, 0);
