+++ views/front-page-components.html
@@ -1,0 +1,5 @@
+<h1>Front Page: Components</h1>
+
+<div id="charts" class="o-grid-container front-page">
+
+</div>
