--- views/front-page-ctr.html
@@ -1,24 +1,0 @@
-<h1>Front Page: CTR</h1>
-
-<div id="charts" class="o-grid-container front-page">
-	<div class="front-page__metrics o-grid-row">
-		<a href="#front-page-users-chart" class="front-page__metric front-page__metric--small js-front-page-users" data-o-grid-colspan="2"></a>
-		<a href="#front-page-views-chart" class="front-page__metric front-page__metric--small js-front-page-views" data-o-grid-colspan="2"></a>
-		<a href="#front-page-ctr-chart" class="front-page__metric js-front-page-ctr" data-o-grid-colspan="4"></a>
-		<a href="#front-page-clicks-per-user-chart" class="front-page__metric js-front-page-clicks-per-user" data-o-grid-colspan="4"></a>
-	</div>
-
-	<div class="front-page__charts o-grid-row">
-		<span id="front-page-users-chart" class="front-page__chart-anchor"></span>
-		<div  class="front-page__chart js-front-page-users-chart" data-o-grid-colspan="12"></div>
-
-		<span id="front-page-views-chart" class="front-page__chart-anchor"></span>
-		<div class="front-page__chart js-front-page-views-chart" data-o-grid-colspan="12"></div>
-
-		<span id="front-page-ctr-chart" class="front-page__chart-anchor"></span>
-		<div class="front-page__chart js-front-page-ctr-chart" data-o-grid-colspan="12"></div>
-
-		<span id="front-page-clicks-per-user-chart" class="front-page__chart-anchor"></span>
-		<div class="front-page__chart js-front-page-clicks-per-user-chart" data-o-grid-colspan="12"></div>
-	</div>
-</div>
